fix(go-core): centralize identity metadata
